+++ lib/src/signup/signup_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_google_maps_webservices/places.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:wash_wow/src/login/login_screen.dart';
-import 'package:wash_wow/src/services/auth_service.dart';
-import 'package:wash_wow/src/services/google_map_service.dart';
+import 'package:wash_wow/src/utility/auth_service.dart';
+import 'package:wash_wow/src/utility/google_map_service.dart';
 
 class SignupScreen extends StatefulWidget {
   @override
