+++ lib/src/utility/auth_service.dart
