+++ lib/src/utility/extension/string_extension.dart
