+++ lib/src/utility/model/store_details.dart
